templates rework
