+++ grafana-plugin/src/components/Integrations/IntegrationBlock.module.scss
