+++ grafana-plugin/src/components/Integrations/IntegrationBlock.tsx
