+++ grafana-plugin/src/components/Integrations/IntegrationBlockItem.module.scss
