+++ grafana-plugin/src/components/Integrations/IntegrationBlockItem.tsx
