+++ grafana-plugin/src/components/Integrations/IntegrationTemplateBlock.module.scss
