+++ grafana-plugin/src/components/Integrations/IntegrationTemplateBlock.tsx
