+++ grafana-plugin/src/pages/integration_2/Integration2.helper.ts
